+++ src/stores/GuildChannelStore.js
@@ -1,9 +1,8 @@
-const Collection = require('../util/Collection');
 const Channel = require('../structures/Channel');
 const { ChannelTypes } = require('../util/Constants');
 const DataStore = require('./DataStore');
 const GuildChannel = require('../structures/GuildChannel');
-const Permissions = require('../util/Permissions');
+const resolvePermissions = require('../structures/shared/resolvePermissions');
 
 /**
  * Stores guild channels.
@@ -24,9 +23,9 @@ class GuildChannelStore extends DataStore {
 
   /**
    * Can be used to overwrite permissions when creating a channel.
-   * @typedef {Object} ChannelCreationOverwrites
-   * @property {PermissionResolvable[]|number} [allow] The permissions to allow
-   * @property {PermissionResolvable[]|number} [deny] The permissions to deny
+   * @typedef {Object} PermissionOverwriteOptions
+   * @property {PermissionResolvable[]|number} [allowed] The permissions to allow
+   * @property {PermissionResolvable[]|number} [denied] The permissions to deny
    * @property {RoleResolvable|UserResolvable} id ID of the role or member this overwrite is for
    */
 
@@ -39,7 +38,7 @@ class GuildChannelStore extends DataStore {
    * @param {number} [options.bitrate] Bitrate of the new channel in bits (only voice)
    * @param {number} [options.userLimit] Maximum amount of users allowed in the new channel (only voice)
    * @param {ChannelResolvable} [options.parent] Parent of the new channel
-   * @param {Array<PermissionOverwrites|ChannelCreationOverwrites>} [options.overwrites] Permission overwrites
+   * @param {Array<PermissionOverwrites|PermissionOverwriteOptions>} [options.overwrites] Permission overwrites
    * @param {string} [options.reason] Reason for creating the channel
    * @returns {Promise<GuildChannel>}
    * @example
@@ -47,33 +46,19 @@ class GuildChannelStore extends DataStore {
    * guild.channels.create('new-general', { reason: 'Needed a cool new channel' })
    *   .then(console.log)
    *   .catch(console.error);
+   * @example
+   * // Create a new channel with overwrites
+   * guild.channels.create('new-voice', {
+   *   type: 'voice',
+   *   overwrites: [
+   *      {
+   *        id: message.author.id,
+   *        denied: ['VIEW_CHANNEL'],
+   *     },
+   *   ],
+   * })
    */
   create(name, { type, nsfw, bitrate, userLimit, parent, overwrites, reason } = {}) {
-    if (overwrites instanceof Collection || overwrites instanceof Array) {
-      overwrites = overwrites.map(overwrite => {
-        let allow = overwrite.allow || (overwrite.allowed ? overwrite.allowed.bitfield : 0);
-        let deny = overwrite.deny || (overwrite.denied ? overwrite.denied.bitfield : 0);
-        if (allow instanceof Array) allow = Permissions.resolve(allow);
-        if (deny instanceof Array) deny = Permissions.resolve(deny);
-
-        const role = this.guild.roles.resolve(overwrite.id);
-        if (role) {
-          overwrite.id = role.id;
-          overwrite.type = 'role';
-        } else {
-          overwrite.id = this.client.users.resolveID(overwrite.id);
-          overwrite.type = 'member';
-        }
-
-        return {
-          allow,
-          deny,
-          type: overwrite.type,
-          id: overwrite.id,
-        };
-      });
-    }
-
     if (parent) parent = this.client.channels.resolveID(parent);
     return this.client.api.guilds(this.guild.id).channels.post({
       data: {
@@ -83,7 +68,7 @@ class GuildChannelStore extends DataStore {
         bitrate,
         user_limit: userLimit,
         parent_id: parent,
-        permission_overwrites: overwrites,
+        permission_overwrites: resolvePermissions.call(this, overwrites),
       },
       reason,
     }).then(data => this.client.actions.ChannelCreate.handle(data).channel);
